refactor(types): take benefit of components
This commit makes use of components to create anonymized components for well separated types.

BREAKING CHANGE: This commit may break uses of anonymous types though it is not recommended, it may
be the case in some situations where one is unable to change the Open API file to fit its needs.

concerns #4 fix #1